Fixed a bug in the code for the forward computation in modules/network_to_softmax_network.py in case minimal horizontal padding was used. The restructuring of the activations into a 2D structure with dimensions: number of activations, number of channels was not done correctly. Because of this the results when training with the minimal horizontal padding were suboptimal.
Added code for last minute padding. This is used when
the proper MDLSTM as opposed to the block-MDLSTM is
being used. In this case, considerable savings can be
made by padding seperately for every batch rather than
uniformly to the maximal dimensions that appear in
the entire training data.

Also improved the structure of the code, extracting
functions in very large methods in modules/network_to_softmax_network.py
and extracting a separate class  modules/module_io_structuring.py
for the specific structuring of input and output to modules in case
of block-chunking, as used for efficient computation with minimal
padding with block_mdlstm.

	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   data_preprocessing/last_minute_padding.py
	modified:   data_preprocessing/load_mnist.py
	new file:   modules/module_io_structuring.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py
	new file:   util/test_tensor_operations.py