Added a combination of gradient clamping (gradient correction during the backward pass), gradient clipping by value and gradient clipping by norm.
The gradient grows explosively with deep MDLSTMs it seems.
Gradient clamping can reduce the growth considerably during
backward computation but alone still appears not to manage
to assure some gradient components are not becoming infinity
anyway. Even with very small clamping value ranges the latter
still happens.

Therefore, a second clipping, based on value, brings the gradients
back to a normal range. Finally, norm-based clipping is done
as a last step on the already gradient-clamping and gradient value
clipping corrected gradients.

This may seems like overkill, and it definitely feels like it,
but it seems to be working.
The gradient value clipping fixes the persistent norm-becomes-nan
problem, which gradient norm clipping by itself could not overcome.
The idea of gradient clamping is to start correcting the gradients
early on, rather than letting them explode completely and then being
forced to reset them completely by gradient clipping on value,
In this way more useful information can be preserved, by curbing
the amount of gradient explosion so to speak.

	modified:   modules/gradient_clipping.py
	modified:   modules/inside_model_gradient_clipping.py
	modified:   modules/optim.py
	modified:   modules/opts.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py