Changed the clamping bounds to values used in https://github.com/t-vi/pytorch-tvmisc/blob/master/misc/graves_handwriting_generation.ipynb A base clamping bound of 10 is used for MDLSTM internal gradient clamping. For the linear layer in  modules/network_to_softmax_network.py a higher bound of 100 is used. Which bounds are good seems to be very much an emprirical matter, but these bounds seem to be working at least very well in the beginning of learning (the loss quicly decreases from 40 to 30 etc...)
TODO: One thing that is unclear is why global gradient-value clipping would
still be nescessary when the model-internal gradient clamping is applied.
It almost seems as if some gradients are still missed and hence become infinity,
whichi is why the model global gradient clamping is needed to put them back
to normal values. But actually this should not be happening. Model internal
gradient clamping, optionally combined with global norm-based gradient clamping
ought to be enough, but for some reason it is not, as in that setup the
norm still gives nan, meaning probably some gradients are becoming infinity
despity the model internal gradient clamping...
Try to find out when which gradients become infinity, so that the ugly
double gradient-value clamping/clipping (both inside the model and global)
can be done away with.

	modified:   modules/block_strided_convolution.py
	modified:   modules/gradient_clipping.py
	modified:   modules/inside_model_gradient_clipping.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/opts.py
	modified:   modules/parallel_multiple_input_convolutions_computation.py
	modified:   modules/parallel_multiple_state_weightings_computation.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py