Nearly completely fixed the problems with nan gradient norms when performing only gradient clamping and norm-based gradient clipping and no additional global value-based gradient clipping. The fix is to apply gradient clamping even for simple multiplications, e.g. when computing the multiple of a state and a gate output. Appearantly even these multiplications can give rise to exploding gradients when not clamped.
For some reason at the start of training there is still often an
example with nan gradient-norm value. This is odd, as it seems that
gradient clamping is now added everywhere, including throughout the
MDLSTM, in the block_strided_convolution and in the network_to_softmax
layer. While the current situation is already a big improvment, and
probably workable for learning, one would find the cause of and then
eliminate these last nans as well.

	new file:   modules/find_bad_gradients.py
	modified:   modules/inside_model_gradient_clipping.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/optim.py
	modified:   modules/parallel_multiple_input_convolutions_computation.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py