Improved the debugging possibilities for finding the cause of the problem with exploding gradients in MDLSTM. First, changed the find_bad_gradients code to work on the handwriting recognition model. For some reason, it was failing on assertions in the code at first, modiefied these so that the code no longer crashes when building the bad gradients graph.
Second, added code to print all gradients after the backwards call.
This revealed that indeed there are nan-gradients.

Third, changed the gradient clamping hooks to register the
name of the variable they are clamping and output the input
before clamping and output after clamping. The results suggest
that the nans appear first in variables in MDLSTM that are
downstream from convolution modules.

TODO:

This raises a quesion: can the exploding gradients result from
the convolution modules themselves (which are not clamped), and
if so is there a way to clamp the gradients of these modules
to avoid exploding gradients for them?
This will be explored next.