Fixed an implementation error in MDLSTM. The values of hidden and memory states that do not correspond to cells in the original image, before skewing it for the convolution computation trick, should be masked / reset to zero before adding them to the activations or saving them as memory/hidden states for the next column in the MDLSTM computation. This was not done, and hence the cells in the different columns were obtaining input from essentially invalid hidden and memory states, taken from invalid predecessor states.
This problem has now been addressed, by computing a binary mask of the same
size as the skewed image, and with zeros in cells that are invalid and ones
in cells that correspond to valid cells in the original image. The columns
of this mask are then used in the MDLSTM column by column computation
to mask for each new memory state and hidden state column that is being
computed the cells that are invalid, essentially setting them all to zeros.

This all seems to work, learning still goes OK on th MNIST data it seems.

TODO:

Unfortunately this fix still does not seem to much help the problem with
the exploding gradients on the IAM words database, for which the input
images are much bigger. In other words, the quest for finding a fix
for the escaping nan gradients must continue.

	new file:   modules/gradient_clamped_module.py
	modified:   modules/inside_model_gradient_clipping.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/parallel_multiple_input_convolutions_computation.py
	modified:   modules/parallel_multiple_state_weightings_computation.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py
	modified:   util/image_input_transformer.py
	modified:   util/tensor_utils.py