After a very long search found the cause of the problems with nans in the gradient. This link was useful in finally getting the right insight: https://github.com/pytorch/pytorch/issues/4132
"Key to avoid nan

Your code should't generat any inf in forward, which often produce by torch.log(0) and x/[0, ]
That means 0 should be filtered before do torch.log(x) and x/div
"

As it turned out, it was not clear that my forward code was causing infinity.
However, putting some prints on the maximum values of tensors in the
MDLSTM computation revealed a clear problem:

1. Earlier it was revealed that "output_gate_weighted_states_plus_input"
was the first place were nan gradients were observed. This was found by
registering a gradient-clamping backward hook on this variable in the function
"compute_multi_dimensional_lstm_one_direction" in multi_dimensional_lstm.py.
This backward hook checked for bad gradients (including nans) and raised
a runtime error once it found one.

2. Second, I observed that in fact none of the predecessor gradients that
were computed before this gradient (the gradient from the convolution layer
after the first MDLSTM layer for example) seemsed to be problematic.
This seemed odd, as it was hard to imagine how the computation of the gradient
of a sigmoid function could lead to problems.

3. Adding information about the layer index and column number to the bad-gradient
identifying hook, revealed that the problem appeared in the last column of the
first MDLSTM layer. That is, at the start of the backward pass for that layer.

4. At the point I was almost stuck with this bug, I decided to take the point of
problematic values in the forward computation itself more serious. So I added
printouts of the maximum values of tensors. It turned out then that the
variables "output_gate_weighted_states_plus_input" and at the root of that
"new_memory_state" were growing unboundedly, with a growth rate of approximately 2
or less, per column iteration in the MDLSTM forward computation, obtaining values
greater than e.g. 1.4322e+10 after a while.

I realized, that new_memory_state would not be gated anymore, but directly saved
for the next column iteration. Hence it seemed to make sense that the
old computation:

"
new_memory_state = input_and_input_gate_combined + \
                forget_gate_two_activation_multiplied_with_previous_memory_state + \
                forget_gate_one_activation_multiplied_with_previous_memory_state # + \
                # forget_gates_combined_activation_multiplied_with_previous_memory_state \
"

could indeed cause the memory state to grow over time, since it effectively replaced new_memory_state
with two functions of the old memory state in addition to the function of the input. Importantly,
both  "forget_gate_two_activation_multiplied_with_previous_memory_state"  and
"forget_gate_one_activation_multiplied_with_previous_memory_state" in the above formula have the
potential to become nearly as large as memory_state, as the activity of their gating component
(memory gate 1, memory gate 2) go to 1.

A simple fix that is added in this commit, and that indeed stabilizes the values is:

"
            new_memory_state = input_and_input_gate_combined + \
                               0.5 * forget_gate_two_activation_multiplied_with_previous_memory_state + \
                               0.5 * forget_gate_one_activation_multiplied_with_previous_memory_state  # + \
"

This fix also removes the need to do gradient clamping during backward computation, now norm-based gradient clipping,
which is computationally cheaper and has nicer theoretical properties (preserving the direction of the original gradient,
which value-based clamping/clipping cannot achieve).

Note however, that another way to fix the problem would be to initialize the foget gates differently, which are
now initialized to 1. While the latter may be advisible for single-dimensional LSTM, it may lead to unnescessary
instability for MultiDimensionalLSTM.

TODO: Next I will look if the scaling of the components by 0.5 can be gotten rid of by changing the bias to the forget
gates instead.

	modified:   modules/block_strided_convolution.py
	modified:   modules/gradient_clamped_module.py
	modified:   modules/inside_model_gradient_clipping.py
	modified:   modules/mdlstm_layer_block_strided_convolution_layer_pair.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/parallel_multiple_state_weightings_computation.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py
	modified:   util/tensor_utils.py