Added the option to already do the conversion to float during data loading. This is somewhat faster, and especially in case of iam_words, doesn't require an unaffordable amount of extra memory.
TODO: With iam_words, in the new implementation with MDLSTM executed with
the entire image rectangle as input, rather than blocks, the amount of
padding is still large. Even using last minute padding, adding minimal
padding within each batch seperately, the amount of padding for batches
still comes to something like 70-80%.

TODO: One unexplored strategy is to sort the examples within a batch by
size, and then split into two or more groups with similar size.
For the smaller group, the amount of padding required will then be less.
This can probably be optimized so that even despite a decrease in parallellism,
still time is saved, by significantly reducing the total amount of
computation on padding.

	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py