Further implemented the MDLSTM examples packing. Contrary to what was written in the last commit message, full examples packing and unpacking is necessary for every MDLSTM layer and also block chunking and de-chunking is necessary for every block-strided convolution layer. The packing seems to work now, but there are still some issues.
DataParallel does not work with the new packing implementation, since
it cannot split lists of examples over multiple GPUs in the standard
implementation.

TODO: A custom version of DataParallel must be made with custom
scatter/gather functions, to split the example lists and forward
function list results over multiple GPUs. This should be possible
(some people in the pytorch forums did it for ordered dictionary),
albeit complicated. Right now, packing only works with one GPU,
making it slower.

Another note:

Fixed a bug in the usage of the mask in the MDLSTM forward function,
in multi_dimensional_lstm.py.
The mask was used to mask the results of the convolution of the hidden
and memory states. This could be ok, but the bug was that the mask
of the current column was used (whereas the mask of the previous
column should be used, which corresponds to the previous memory/hidden
states). Changed the way this is implemented by just masking the
hidden/activation/memory states by the current column mask, before
ending the loop of the current column in the MDLSTM forward function.
This is simpler and arguably cleaner: we are not interferring with the
MDLSTM computation itself, just setting the hidden and memory states
to zero for cels that are masked, just as they are zero in the beginning
for the first cells that are computed.

	modified:   ../data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   ../data_preprocessing/load_mnist.py
	modified:   block_strided_convolution.py
	modified:   mdlstm_examples_packing.py
	modified:   mdlstm_layer_block_strided_convolution_layer_pair.py
	modified:   module_io_structuring.py
	modified:   multi_dimensional_lstm.py
	modified:   multi_dimensional_lstm_layer_pair_stacking.py
	modified:   network_to_softmax_network.py
	modified:   parallel_multiple_state_weightings_computation.py
	modified:   size_two_dimensional.py
	modified:   train_multi_dimensional_rnn_ctc.py
	modified:   ../util/image_input_transformer.py
	modified:   ../util/image_visualization.py
	modified:   ../util/tensor_list_chunking.py
	modified:   ../util/tensor_utils.py