Implemented a custom version of data_parallel that works for modules that have inputs and outputs in the form of lists. The normal data_parallel implementation cannot parallelize input lists over multiple gpus. The new implementation splits the lists into sublists, similar to the way the normal data_parallel splits tensors into chunks, and then computes the results for these sublists on different gpus.
How exactly to implement this, that is how to customize data_parallel in
a way that works, required quite a lot of search. Whether at first it appeared
that the file _functions.py, which implements Scatter and Gather classes that
extend torch.autograd.Funcion needed to be customized, later it turned out
this was not the way to go. Directly customizing the functions in "comm.py"
which is originalally found in "torch.cuda.comm" was what eventually worked.
These customized functions are then called at the right places inside
the customized version of "scatter_gather.py".

Whereas the gains could have been bigger (the whole approach also adds some
extra overhead), at least with iam_words it does seem to pay of.
The main gain is that using all GPUs, the amount of memory used per GPU
is less, so the batch size can be increased (e.g. to 256),
and this then gives faster performance.
This goes in combination with the earlier optimization of packing examples
horizontally and vertically, which while also adding quite some overhead,
drastically reduces the amount of padding and hence the amount of
computation and memory used - again enabling increase of the batch size
used.

TODO: Things can be further optimized for multi_directional computation,
using the same example packing for the different directions, and
packing the tensors for the different directions all in the same packed
tensor.

	new file:   custom_data_parallel/__init__.py
	new file:   custom_data_parallel/_functions.py
	new file:   custom_data_parallel/_functions_lists.py
	new file:   custom_data_parallel/comm_list.py
	new file:   custom_data_parallel/data_parallel.py
	new file:   custom_data_parallel/scatter_gather.py
	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   util/utils.py

"#	.output-iam-words_stable_vocabulary_Adam_lr0.0001-batch_size_32_80_epochs-fixed_data_permutation-minimal-padding-hight-and-width.txt.swp