Added first of all a bug-fix to the method taking care of the extraction of unskewed activations for example rows :extract_unskewed_activations_packed_examples_row":
-        activations_unskewed = activations_as_tensor[:, :, 0, 0:original_image_columns]
+        # The first row of this example row is at vertical index first_row_index (not 0) !!!
+        activations_unskewed = activations_as_tensor[:, :, first_row_index, 0:original_image_columns]

It turned out the wrong row index was used for the first row of the activation rows corresponding
to the example row. Consequently, the same (first) row of activations was constantly used, a pretty
serious bug.

Secondly, this fix still did not resolve the problem that the MDLSTM activations were not the same
with and without packing. After digging deep in the code, and determining that although the
convolution results for the first hidden and memory states column (which is initialized always to zeros)
is the same for every entry, the final hidden state activations were actually different for the
first element of this column and the rest of the elelements.
Then determined that the reason for this is actually, that the convolution result is added to the
one-down-shifted convolution result, to achieve summation of the two previous states.
Came to the conclusion that the only way to get the same results with examples packing is therefore
to also apply a mask (not for the current input colum but for the previous states,
that are used as convolution input) when computing the convolution results for computing the new hidden
states and memory states activations. With this done, combined with the other fix,
the results for MDLSTM with or without packing become indeed the same, at least for a bunch of
test examples, from very simple to slightly more complex.

TODO: The fixed implementation for examples packing should again be tested to see if it also performs
as well as the non-packing implementation on MNIST variable length and on IAM word recognition.

	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/parallel_multiple_state_weightings_computation.py
	modified:   tests/multi_dimensional_lstm_test.py