Added a multi-directional variant of multi-dimensional LSTM that uses a parallel_multiple_inputs_convolution that is parallelized over : 1. The different hidden states and memory states input weighting 2. The different directions That is, given enough parallel GPU power (or a small enough batch size), the multi-directional MDLSTM implementation can compute the whole MDLSTM column computation almost entirely in parallel. This is as opposed to computing the MDLSTM first for one direction, then for the next, until 4 directions are done, one after the other.
This is important, because whereas more GPU usage /parallelization can always be used
by increasing the batch-size for a poorly-parallelized implementation,
a too large batch-size is not good for training handwriting recognition
models. Furthermore, because loss is normalized by the batch size
before back-propagation, it means that the summed magnitude of updates for the
summed steps of one epoch is also to be expected to be effectively lower
when the batch-size is bigger. These are all reasons to want to use a small
batch size, while retaining as much GPU usage (i.e. parallelization) as
possible.

Because it is easy to introduce bugs in the multi-directional implementation,
and indeed bugs were there in the beginning, a test was added to see if
the output of four one-directional MDLSTM correspond to the corresponding
outputs of a single 4-directional MDLSTM. To test this properly, the
weights between the 1-directional MDLSTMs and the 4-directional MDLSTM
must correspond. Therefore, methods to create 1-directional MDLSTMs
using the corresponding weights of a 4-directional MDLSTM were also
made. The test checks that the outputs correspond for increasingly
complex artificial input tensors.

TODO: What is left for the 4-directional MDLSTM implementation is
adapting the block-strided convolution to process the lists of
4-directional MDLSTM activation tensors, with the activations of
the four directions concatenated along the batch dimension
(the first dimension).

	modified:   modules/block_strided_convolution.py
	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/mdlstm_layer_block_strided_convolution_layer_pair.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/parallel_multiple_input_convolutions_computation.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   tests/multi_dimensional_lstm_test.py
	new file:   tests/multi_directional_mdlstm_test.py
	new file:   tests/test_conv1d.py