Restored the computation of the multiple input convolutions for different directions with seperate convolutions. It appears that using grouping here can cause errors with CUDA of the form
"RuntimeError: CUDNN_STATUS_NOT_SUPPORTED. This error may appear if you passed in a non-contiguous input."
See e.g. https://stackoverflow.com/questions/45276997/cudnn-error-in-backprop-for-big-batches
It is not entirely clear, but it seems plausible that doing the convolutions for the different
input directions seperate is more "safe". As these convolutions can become rather big otherwise,
and this might be what is causing these problems, as they did not appear to occur before.

	modified:   ctc_loss/warp_ctc_loss_interface.py
	modified:   libraries/warp-ctc
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   modules/trainer.py
	modified:   tests/multi_directional_mdlstm_test.py