Attempted a few more optimizations, but nothing really gave further improvements. Adding some timekeeping it became clear that in fact the main computation of the MDLSTM layer takes the most time, up two four times as much as the packing/unpacking. Also, this is inherently hard to parallelize further. The only thing that can be done to increase parallelization is increasing the batch size, but this increases the risk of running out of memory.
TODO: One optimization that could still be explored is a more even
distribution of examples between the different GPUs. However, it is
questionable, if the current, simply random division, is so bad.
But maybe a slightly larger batch-size can be safely used if the
examples are guaranteed to be almost exactly evenly
(in terms of their summed sizes) dividided  between the GPUs.

	modified:   modules/block_strided_convolution.py
	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/parallel_multiple_state_weightings_computation.py
	modified:   modules/trainer.py
	modified:   util/image_input_transformer.py