Added a version/parameterization of the MDLSTM network that starts with only 2 hidden states per direction in the first MDLSTM layer and mostly follows the parametrization for the number of hidden states per layer from the paper "Dropout improves Recurrent Neural Networks for Handwriting Recognition".
Adapted the version of the MDLSTM weights computation
"MultiDirectionalMultiDimensionalLSTMParametersParallelWithSeparateInputConvolution"
to save the input convolution results in a list of lists, rather than a list of
input (convolution result) matrices. The idea is then that these columsn can be
removed from the lists after they are used, to free up memory on the go as
the column-by-column computation takes place in the network.

	modified:   data_preprocessing/iam_database_preprocessing/handwriting_centering_and_clipping.py
	modified:   modules/block_strided_convolution.py
	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/mdlstm_layer_block_strided_convolution_layer_pair.py
	modified:   modules/multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py