Restored the usage of the MDLSTM fully parallel, and for the variable-length MNIST code in modules/train_multi_dimensional_rnn_ctc.py, restored using 4-directional computation, 1-3 digits and usage of packing.
Found out that this was all fine. The real reason for the poor performance
on variable-length MNIST was a too low learning rate (0.0005) set in the
program arguments of the pycharm GUI. With a learning rate of 0.05 the
learning after even just 2 epochs is already substantial, as it was before,
before the "bug" (which was essentially just a wrong learning
rate set in the pycharm GUI program aruments) appeared.

"
>>> evaluate_mdrnn  - output: "8"
reference: "5" --- wrong
>>> evaluate_mdrnn  - output: "3"
reference: "3" --- correct
>>> evaluate_mdrnn  - output: "88"
reference: "38" --- wrong
Accuracy of the network on the 5035 test inputs: 41.93 % accuracy
Character Error Rate (CER) of the network on the 5035 test inputs: 0.390  CER
Word Error Rate (WER) of the network on the 5035 test inputs: 0.620  WER
"

Conclusion and TODO:

An additional unit test using MNIST variable length digits is needed,
that trains for 2 epochs (4-directional MDSLTM) and checks accuracy is say at least 15 %.
The learning rate should be set in the test. Furthermore: the code should probably be adapted
so that the learning rate will not decrease by default if
“-start_decay_at 10000 -learning_rate_decay 1“
are not set.