Added a test that shows how ctcdecode can be used with a Kenlm language model, and how whether this language model is used (with a non-zero weight) or not used, really changes the decoding results.
TODO: An open question is still what to do with the word space ("|")
symbol during language model training. This symbol is different from
the blank symbol. This raises the question whether the words used
for training the language model should be extended with the "|"
symbol, or else what kind of other trick needs to be used to harmonize
the language model training with the ctc decoding?

	modified:   language_model/kenlm_interface.py
	modified:   modules/evaluator.py
	modified:   tests/test_ctc_decode_with_language_model.py