Added classes for creating a language model for the LOB (Lancaster-Oslo-Bergen) corpus of modern English. First a version of the lines file is created with only the words and more suitable whitespaces and newlines, using the data_preprocessing/monolingual_data_preprocessing/untagged_text_extractor.py class.
Then, using  data_preprocessing/monolingual_data_preprocessing/lines_with_explicit_word_separator_file_creator.py,
optionally explicit word separators can be added to the lines in between the words, to train a language model
that contains these explicit word separators. Alternatively, simply a language model without the explicit word
separators can be used (which cannot account then for the word separators, which is a bit of a problem).

The evaluator class is also adapted, to create a decoder with lanugage model (when this is switched on
via opts.py). If the language model is used, the decoder output must be post-processed, replacing
whitespaces in the output with the word separator symbol "|". Again, this is not ideal, since there can
also be real whitespace characters, although these are rare.

TODO: A word frequency count table should be created and used to replace the non-frequent words with
a special NON_FREQUENT_WORD symbol in the data produced for language model training. This way
a fixed-size vocabulary can be simulated, since the ctcdecode decoder does not have an option to
provide an explicit and separate vocabulary.

	new file:   data_preprocessing/monolingual_data_preprocessing/__init__.py
	new file:   data_preprocessing/monolingual_data_preprocessing/lines_with_explicit_word_separator_file_creator.py
	new file:   data_preprocessing/monolingual_data_preprocessing/untagged_text_extractor.py
	modified:   language_model/kenlm_interface.py
	modified:   modules/evaluator.py
	modified:   modules/opts.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   tests/test_ctc_decode_with_language_model.py