Changed and renamed the method "create_train_loader" => "create_data_loader" in data_preprocessing/padding_strategy.py, adding a third argument "shuffle", with a boolean whether the data should be shuffled or not. For the validation and test-set, shuffling is not suitable, as it adds unnescessary ranomness to the order of the input which makes it harder to reproduce results.
What was discovered hower, is that the evaluation results are not stable
across multiple runs with the same decoder settings and the same test set.

TODO: Find out what factors including the varying input order (fixed in this commit)
cause different results on the test set over multiple repetitions, which should
not occur.

	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   data_preprocessing/load_mnist.py
	new file:   data_preprocessing/monolingual_data_preprocessing/iam_database_fragments_remover.py
	modified:   data_preprocessing/padding_strategy.py
	modified:   util/utils.py