Fixed one bug and one conceptual error in the IAM dataset creation: 	1. The bug was that in the method "IAMLinesDataset".split_specified_train_set_validation_set_and_test_set" 	   the examples that did not belong to either the train, dev or test-set ids-set were still assigned to            the train set.         2. The bug was to use only the "ok" examples in the IAM dataset. Since in the lines.txt file it actually says:
	" notice: if the line could not be properly segmented the transcription and extraction of the whole
                  line should not be affected negatively. "
In other words, all the lines includding the "err" marked lines can be safely used when doing entire-line
training and recognition, and this is what people do in the literature i.e. in
"Dropout improves Recurrent Neural Networks for Handwriting Recognition" (Pham. et. al, 2013) and
"Handwriting Recognition with Large Multidimensional Long Short-Term Memory
Recurrent Neural Networks" (Voigtlander et.al, 2016).

With this fixed, the right sizes for training, validation and testing are obtained when
using the Bluche splits
from http://www.tbluche.com/files/iam-splits/:

>>> number of training examples: 6482
>>> number of development examples: 976
>>> number of test examples: 2915
>>> number of omitted examples: 2980

The fact that some of the examples are omitted seems to be not explicitly reported in these papers, but that is fine,
as long as we can use exactly the same set.

	modified:   data_preprocessing/iam_database_preprocessing/iam_dataset.py
	modified:   data_preprocessing/iam_database_preprocessing/iam_examples_dictionary.py