Improved the initialization of the softmax layer when four independent layers with unique weights, one for each direction are used. These four layers are implemented as one single linear layer with four times as many inputs. But this means that Xavier uniform weight initialization: Wi,j = U(-srqt(6/(m+n), sqrt(6/(m+n)) with for m the network layer fan-in, and n the fan-out will have a value for m that is four times larger than when four separate layers were used. This leads to sqrt(4) = 2 times smaller weights. Hence, as a compensation we re-multiply the weight factor afterwards again with a factor 2 for compensation. This is motivated by the fact that seemingly the final fully connected layer with weight sharing is learning faster in the beginning. This may well be due to the effectively larger weights due to the way the Xavier initialization is done. The now implemented weight correct should remove this difference.
TODO: An open question is still how to implement the bias. Changed
it now to initialize them with zero. This is different from the default,
which initializes them to uniform(-stdv,stdv), but could be better:
https://discuss.pytorch.org/t/how-are-layer-weights-and-biases-initialized-by-default/13073.
It is suggested as a good default in the deep learning textbook.
There it is also mentioned that for the final layer, an initialization that reflects the
marginal probabilities of the outputs could be even better. However, it is not clear that
the marginal probability over the label outputs maps to the desired marginal probability over
the activation outputs, since ctc loss is in between. Probably, the blank output should
get most of the prior probability, since it is described in the literature that the
networks often learn to output blanks often and only peak with a character output at
places where it is very certain enough evidence for a specific character has been collected.
How to map this to deriving from marginal distributions over characters in the training labels
is not clear.