Added implementation of dropout as used in "Dropout improves Recurrent Neural Networks for Handwriting Recognition". The dropout is applied to the output of the MDLSTM layers, and the number of output channels of the MDLSTM layers is increased by a factor 2 (to compensate for dropout with 0.5). There are other ways to apply dropout, for example in "Handwriting Recognition with Large Multidimensional Long Short-Term Meory Recurrent Neural Networks" :
"We use dropout of 0.25 for the forward connections of all convolutional layers,
 MDLSTM layers and the output layer, except for the first
 convolutional layer, where we don’t use dropout as we don’t
 want to drop the single color channel of the input image. "

The way used by Bluche et.al is a bit simpler though, and since their network structure is
mostly replicated it makes more sense to use that way of performing dropout as a start.
A disadvantage of using dropout and doubling all the MDLSTM layer sizes is of course everything
will take twice as long to train. But there is not much to do about that, given the way dropout
is implemented in pytorch, which is simple but not really optimized for performance.

	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/opts.py
	modified:   modules/parallel_multiple_input_convolutions_computation.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py